Move models to legacy,rename models_name to models, refine README
