+++ fluid/PaddleCV/image_classification/legacy/models/fast_imagenet.py
