+++ fluid/PaddleCV/image_classification/legacy/models/mobilenet.py
@@ -32,8 +32,7 @@ class MobileNet():
             channels=3,
             num_filters=int(32 * scale),
             stride=2,
-            padding=1,
-            name="conv1")
+            padding=1)
 
         # 56x56
         input = self.depthwise_separable(
@@ -42,8 +41,7 @@ class MobileNet():
             num_filters2=64,
             num_groups=32,
             stride=1,
-            scale=scale,
-            name="conv2_1")
+            scale=scale)
 
         input = self.depthwise_separable(
             input,
@@ -51,8 +49,7 @@ class MobileNet():
             num_filters2=128,
             num_groups=64,
             stride=2,
-            scale=scale,
-            name="conv2_2")
+            scale=scale)
 
         # 28x28
         input = self.depthwise_separable(
@@ -61,8 +58,7 @@ class MobileNet():
             num_filters2=128,
             num_groups=128,
             stride=1,
-            scale=scale,
-            name="conv3_1")
+            scale=scale)
 
         input = self.depthwise_separable(
             input,
@@ -70,8 +66,7 @@ class MobileNet():
             num_filters2=256,
             num_groups=128,
             stride=2,
-            scale=scale,
-            name="conv3_2")
+            scale=scale)
 
         # 14x14
         input = self.depthwise_separable(
@@ -80,8 +75,7 @@ class MobileNet():
             num_filters2=256,
             num_groups=256,
             stride=1,
-            scale=scale,
-            name="conv4_1")
+            scale=scale)
 
         input = self.depthwise_separable(
             input,
@@ -89,8 +83,7 @@ class MobileNet():
             num_filters2=512,
             num_groups=256,
             stride=2,
-            scale=scale,
-            name="conv4_2")
+            scale=scale)
 
         # 14x14
         for i in range(5):
@@ -100,8 +93,7 @@ class MobileNet():
                 num_filters2=512,
                 num_groups=512,
                 stride=1,
-                scale=scale,
-                name="conv5" + "_" + str(i + 1))
+                scale=scale)
         # 7x7
         input = self.depthwise_separable(
             input,
@@ -109,8 +101,7 @@ class MobileNet():
             num_filters2=1024,
             num_groups=512,
             stride=2,
-            scale=scale,
-            name="conv5_6")
+            scale=scale)
 
         input = self.depthwise_separable(
             input,
@@ -118,8 +109,7 @@ class MobileNet():
             num_filters2=1024,
             num_groups=1024,
             stride=1,
-            scale=scale,
-            name="conv6")
+            scale=scale)
 
         input = fluid.layers.pool2d(
             input=input,
@@ -130,9 +120,7 @@ class MobileNet():
 
         output = fluid.layers.fc(input=input,
                                  size=class_dim,
-                                 param_attr=ParamAttr(
-                                     initializer=MSRA(), name="fc7_weights"),
-                                 bias_attr=ParamAttr(name="fc7_offset"))
+                                 param_attr=ParamAttr(initializer=MSRA()))
         return output
 
     def conv_bn_layer(self,
@@ -144,8 +132,7 @@ class MobileNet():
                       channels=None,
                       num_groups=1,
                       act='relu',
-                      use_cudnn=True,
-                      name=None):
+                      use_cudnn=True):
         conv = fluid.layers.conv2d(
             input=input,
             num_filters=num_filters,
@@ -155,26 +142,12 @@ class MobileNet():
             groups=num_groups,
             act=None,
             use_cudnn=use_cudnn,
-            param_attr=ParamAttr(
-                initializer=MSRA(), name=name + "_weights"),
+            param_attr=ParamAttr(initializer=MSRA()),
             bias_attr=False)
-        bn_name = name + "_bn"
-        return fluid.layers.batch_norm(
-            input=conv,
-            act=act,
-            param_attr=ParamAttr(name=bn_name + "_scale"),
-            bias_attr=ParamAttr(name=bn_name + "_offset"),
-            moving_mean_name=bn_name + '_mean',
-            moving_variance_name=bn_name + '_variance')
+        return fluid.layers.batch_norm(input=conv, act=act)
 
-    def depthwise_separable(self,
-                            input,
-                            num_filters1,
-                            num_filters2,
-                            num_groups,
-                            stride,
-                            scale,
-                            name=None):
+    def depthwise_separable(self, input, num_filters1, num_filters2, num_groups,
+                            stride, scale):
         depthwise_conv = self.conv_bn_layer(
             input=input,
             filter_size=3,
@@ -182,14 +155,12 @@ class MobileNet():
             stride=stride,
             padding=1,
             num_groups=int(num_groups * scale),
-            use_cudnn=False,
-            name=name + "_dw")
+            use_cudnn=False)
 
         pointwise_conv = self.conv_bn_layer(
             input=depthwise_conv,
             filter_size=1,
             num_filters=int(num_filters2 * scale),
             stride=1,
-            padding=0,
-            name=name + "_sep")
+            padding=0)
         return pointwise_conv
