+++ fluid/PaddleCV/image_classification/legacy/models/mobilenet_v2.py
@@ -36,40 +36,33 @@ class MobileNetV2():
             (6, 320, 1, 1),
         ]
 
-        #conv1 
         input = self.conv_bn_layer(
             input,
             num_filters=int(32 * scale),
             filter_size=3,
             stride=2,
             padding=1,
-            if_act=True,
-            name='conv1_1')
+            if_act=True)
 
-        # bottleneck sequences
-        i = 1
         in_c = int(32 * scale)
         for layer_setting in bottleneck_params_list:
             t, c, n, s = layer_setting
-            i += 1
             input = self.invresi_blocks(
                 input=input,
                 in_c=in_c,
                 t=t,
                 c=int(c * scale),
                 n=n,
-                s=s,
-                name='conv' + str(i))
+                s=s, )
             in_c = int(c * scale)
-        #last_conv
+
         input = self.conv_bn_layer(
             input=input,
             num_filters=int(1280 * scale) if scale > 1.0 else 1280,
             filter_size=1,
             stride=1,
             padding=0,
-            if_act=True,
-            name='conv9')
+            if_act=True)
 
         input = fluid.layers.pool2d(
             input=input,
@@ -80,8 +73,7 @@ class MobileNetV2():
 
         output = fluid.layers.fc(input=input,
                                  size=class_dim,
-                                 param_attr=ParamAttr(name='fc10_weights'),
-                                 bias_attr=ParamAttr(name='fc10_offset'))
+                                 param_attr=ParamAttr(initializer=MSRA()))
         return output
 
     def conv_bn_layer(self,
@@ -92,9 +84,8 @@ class MobileNetV2():
                       padding,
                       channels=None,
                       num_groups=1,
-                      if_act=True,
-                      name=None,
-                      use_cudnn=True):
+                      use_cudnn=True,
+                      if_act=True):
         conv = fluid.layers.conv2d(
             input=input,
             num_filters=num_filters,
@@ -104,15 +95,9 @@ class MobileNetV2():
             groups=num_groups,
             act=None,
             use_cudnn=use_cudnn,
-            param_attr=ParamAttr(name=name + '_weights'),
+            param_attr=ParamAttr(initializer=MSRA()),
             bias_attr=False)
-        bn_name = name + '_bn'
-        bn = fluid.layers.batch_norm(
-            input=conv,
-            param_attr=ParamAttr(name=bn_name + "_scale"),
-            bias_attr=ParamAttr(name=bn_name + "_offset"),
-            moving_mean_name=bn_name + '_mean',
-            moving_variance_name=bn_name + '_variance')
+        bn = fluid.layers.batch_norm(input=conv)
         if if_act:
             return fluid.layers.relu6(bn)
         else:
@@ -121,18 +106,10 @@ class MobileNetV2():
     def shortcut(self, input, data_residual):
         return fluid.layers.elementwise_add(input, data_residual)
 
-    def inverted_residual_unit(self,
-                               input,
-                               num_in_filter,
-                               num_filters,
-                               ifshortcut,
-                               stride,
-                               filter_size,
-                               padding,
-                               expansion_factor,
-                               name=None):
+    def inverted_residual_unit(self, input, num_in_filter, num_filters,
+                               ifshortcut, stride, filter_size, padding,
+                               expansion_factor):
         num_expfilter = int(round(num_in_filter * expansion_factor))
-
         channel_expand = self.conv_bn_layer(
             input=input,
             num_filters=num_expfilter,
@@ -140,9 +117,7 @@ class MobileNetV2():
             stride=1,
             padding=0,
             num_groups=1,
-            if_act=True,
-            name=name + '_expand')
-
+            if_act=True)
         bottleneck_conv = self.conv_bn_layer(
             input=channel_expand,
             num_filters=num_expfilter,
@@ -151,9 +126,7 @@ class MobileNetV2():
             padding=padding,
             num_groups=num_expfilter,
             if_act=True,
-            name=name + '_dwise',
             use_cudnn=False)
-
         linear_out = self.conv_bn_layer(
             input=bottleneck_conv,
             num_filters=num_filters,
@@ -161,15 +134,14 @@ class MobileNetV2():
             stride=1,
             padding=0,
             num_groups=1,
-            if_act=False,
-            name=name + '_linear')
+            if_act=False)
         if ifshortcut:
             out = self.shortcut(input=input, data_residual=linear_out)
             return out
         else:
             return linear_out
 
-    def invresi_blocks(self, input, in_c, t, c, n, s, name=None):
+    def invresi_blocks(self, input, in_c, t, c, n, s):
         first_block = self.inverted_residual_unit(
             input=input,
             num_in_filter=in_c,
@@ -178,8 +150,7 @@ class MobileNetV2():
             stride=s,
             filter_size=3,
             padding=1,
-            expansion_factor=t,
-            name=name + '_1')
+            expansion_factor=t)
 
         last_residual_block = first_block
         last_c = c
@@ -193,6 +164,5 @@ class MobileNetV2():
                 stride=1,
                 filter_size=3,
                 padding=1,
-                expansion_factor=t,
-                name=name + '_' + str(i + 1))
+                expansion_factor=t)
         return last_residual_block
