+++ fluid/PaddleCV/image_classification/legacy/models/resnet.py
@@ -4,7 +4,6 @@ from __future__ import print_function
 import paddle
 import paddle.fluid as fluid
 import math
-from paddle.fluid.param_attr import ParamAttr
 
 __all__ = ["ResNet", "ResNet50", "ResNet101", "ResNet152"]
 
@@ -41,12 +40,7 @@ class ResNet():
         num_filters = [64, 128, 256, 512]
 
         conv = self.conv_bn_layer(
-            input=input,
-            num_filters=64,
-            filter_size=7,
-            stride=2,
-            act='relu',
-            name="conv1")
+            input=input, num_filters=64, filter_size=7, stride=2, act='relu')
         conv = fluid.layers.pool2d(
             input=conv,
             pool_size=3,
@@ -56,18 +50,10 @@ class ResNet():
 
         for block in range(len(depth)):
             for i in range(depth[block]):
-                if layers in [101, 152] and block == 2:
-                    if i == 0:
-                        conv_name = "res" + str(block + 2) + "a"
-                    else:
-                        conv_name = "res" + str(block + 2) + "b" + str(i)
-                else:
-                    conv_name = "res" + str(block + 2) + chr(97 + i)
                 conv = self.bottleneck_block(
                     input=conv,
                     num_filters=num_filters[block],
-                    stride=2 if i == 0 and block != 0 else 1,
-                    name=conv_name)
+                    stride=2 if i == 0 and block != 0 else 1)
 
         pool = fluid.layers.pool2d(
             input=conv, pool_size=7, pool_type='avg', global_pooling=True)
@@ -85,8 +71,7 @@ class ResNet():
                       filter_size,
                       stride=1,
                       groups=1,
-                      act=None,
-                      name=None):
+                      act=None):
         conv = fluid.layers.conv2d(
             input=input,
             num_filters=num_filters,
@@ -95,55 +80,31 @@ class ResNet():
             padding=(filter_size - 1) // 2,
             groups=groups,
             act=None,
-            param_attr=ParamAttr(name=name + "_weights"),
-            bias_attr=False,
-            name=name + '.conv2d.output.1')
-        if name == "conv1":
-            bn_name = "bn_" + name
-        else:
-            bn_name = "bn" + name[3:]
-        return fluid.layers.batch_norm(
-            input=conv,
-            act=act,
-            name=bn_name + '.output.1',
-            param_attr=ParamAttr(name=bn_name + '_scale'),
-            bias_attr=ParamAttr(bn_name + '_offset'),
-            moving_mean_name=bn_name + '_mean',
-            moving_variance_name=bn_name + '_variance', )
+            bias_attr=False)
+        return fluid.layers.batch_norm(input=conv, act=act)
 
-    def shortcut(self, input, ch_out, stride, name):
+    def shortcut(self, input, ch_out, stride):
         ch_in = input.shape[1]
         if ch_in != ch_out or stride != 1:
-            return self.conv_bn_layer(input, ch_out, 1, stride, name=name)
+            return self.conv_bn_layer(input, ch_out, 1, stride)
         else:
             return input
 
-    def bottleneck_block(self, input, num_filters, stride, name):
+    def bottleneck_block(self, input, num_filters, stride):
         conv0 = self.conv_bn_layer(
-            input=input,
-            num_filters=num_filters,
-            filter_size=1,
-            act='relu',
-            name=name + "_branch2a")
+            input=input, num_filters=num_filters, filter_size=1, act='relu')
         conv1 = self.conv_bn_layer(
             input=conv0,
             num_filters=num_filters,
             filter_size=3,
             stride=stride,
-            act='relu',
-            name=name + "_branch2b")
+            act='relu')
         conv2 = self.conv_bn_layer(
-            input=conv1,
-            num_filters=num_filters * 4,
-            filter_size=1,
-            act=None,
-            name=name + "_branch2c")
+            input=conv1, num_filters=num_filters * 4, filter_size=1, act=None)
 
-        short = self.shortcut(
-            input, num_filters * 4, stride, name=name + "_branch1")
+        short = self.shortcut(input, num_filters * 4, stride)
 
-        return fluid.layers.elementwise_add(
-            x=short, y=conv2, act='relu', name=name + ".add.output.5")
+        return fluid.layers.elementwise_add(x=short, y=conv2, act='relu')
 
 
 def ResNet50():
