+++ fluid/PaddleCV/image_classification/legacy/models/resnet_dist.py
