+++ fluid/PaddleCV/image_classification/models/mobilenet.py
@@ -32,7 +32,8 @@ class MobileNet():
             channels=3,
             num_filters=int(32 * scale),
             stride=2,
-            padding=1)
+            padding=1,
+            name="conv1")
 
         # 56x56
         input = self.depthwise_separable(
@@ -41,7 +42,8 @@ class MobileNet():
             num_filters2=64,
             num_groups=32,
             stride=1,
-            scale=scale)
+            scale=scale,
+            name="conv2_1")
 
         input = self.depthwise_separable(
             input,
@@ -49,7 +51,8 @@ class MobileNet():
             num_filters2=128,
             num_groups=64,
             stride=2,
-            scale=scale)
+            scale=scale,
+            name="conv2_2")
 
         # 28x28
         input = self.depthwise_separable(
@@ -58,7 +61,8 @@ class MobileNet():
             num_filters2=128,
             num_groups=128,
             stride=1,
-            scale=scale)
+            scale=scale,
+            name="conv3_1")
 
         input = self.depthwise_separable(
             input,
@@ -66,7 +70,8 @@ class MobileNet():
             num_filters2=256,
             num_groups=128,
             stride=2,
-            scale=scale)
+            scale=scale,
+            name="conv3_2")
 
         # 14x14
         input = self.depthwise_separable(
@@ -75,7 +80,8 @@ class MobileNet():
             num_filters2=256,
             num_groups=256,
             stride=1,
-            scale=scale)
+            scale=scale,
+            name="conv4_1")
 
         input = self.depthwise_separable(
             input,
@@ -83,7 +89,8 @@ class MobileNet():
             num_filters2=512,
             num_groups=256,
             stride=2,
-            scale=scale)
+            scale=scale,
+            name="conv4_2")
 
         # 14x14
         for i in range(5):
@@ -93,7 +100,8 @@ class MobileNet():
                 num_filters2=512,
                 num_groups=512,
                 stride=1,
-                scale=scale)
+                scale=scale,
+                name="conv5" + "_" + str(i + 1))
         # 7x7
         input = self.depthwise_separable(
             input,
@@ -101,7 +109,8 @@ class MobileNet():
             num_filters2=1024,
             num_groups=512,
             stride=2,
-            scale=scale)
+            scale=scale,
+            name="conv5_6")
 
         input = self.depthwise_separable(
             input,
@@ -109,7 +118,8 @@ class MobileNet():
             num_filters2=1024,
             num_groups=1024,
             stride=1,
-            scale=scale)
+            scale=scale,
+            name="conv6")
 
         input = fluid.layers.pool2d(
             input=input,
@@ -120,7 +130,9 @@ class MobileNet():
 
         output = fluid.layers.fc(input=input,
                                  size=class_dim,
-                                 param_attr=ParamAttr(initializer=MSRA()))
+                                 param_attr=ParamAttr(
+                                     initializer=MSRA(), name="fc7_weights"),
+                                 bias_attr=ParamAttr(name="fc7_offset"))
         return output
 
     def conv_bn_layer(self,
@@ -132,7 +144,8 @@ class MobileNet():
                       channels=None,
                       num_groups=1,
                       act='relu',
-                      use_cudnn=True):
+                      use_cudnn=True,
+                      name=None):
         conv = fluid.layers.conv2d(
             input=input,
             num_filters=num_filters,
@@ -142,12 +155,26 @@ class MobileNet():
             groups=num_groups,
             act=None,
             use_cudnn=use_cudnn,
-            param_attr=ParamAttr(initializer=MSRA()),
+            param_attr=ParamAttr(
+                initializer=MSRA(), name=name + "_weights"),
             bias_attr=False)
-        return fluid.layers.batch_norm(input=conv, act=act)
+        bn_name = name + "_bn"
+        return fluid.layers.batch_norm(
+            input=conv,
+            act=act,
+            param_attr=ParamAttr(name=bn_name + "_scale"),
+            bias_attr=ParamAttr(name=bn_name + "_offset"),
+            moving_mean_name=bn_name + '_mean',
+            moving_variance_name=bn_name + '_variance')
 
-    def depthwise_separable(self, input, num_filters1, num_filters2, num_groups,
-                            stride, scale):
+    def depthwise_separable(self,
+                            input,
+                            num_filters1,
+                            num_filters2,
+                            num_groups,
+                            stride,
+                            scale,
+                            name=None):
         depthwise_conv = self.conv_bn_layer(
             input=input,
             filter_size=3,
@@ -155,12 +182,14 @@ class MobileNet():
             stride=stride,
             padding=1,
             num_groups=int(num_groups * scale),
-            use_cudnn=False)
+            use_cudnn=False,
+            name=name + "_dw")
 
         pointwise_conv = self.conv_bn_layer(
             input=depthwise_conv,
             filter_size=1,
             num_filters=int(num_filters2 * scale),
             stride=1,
-            padding=0)
+            padding=0,
+            name=name + "_sep")
         return pointwise_conv
