+++ fluid/PaddleCV/image_classification/models/mobilenet_v2.py
@@ -36,33 +36,40 @@ class MobileNetV2():
             (6, 320, 1, 1),
         ]
 
+        #conv1 
         input = self.conv_bn_layer(
             input,
             num_filters=int(32 * scale),
             filter_size=3,
             stride=2,
             padding=1,
-            if_act=True)
+            if_act=True,
+            name='conv1_1')
 
+        # bottleneck sequences
+        i = 1
         in_c = int(32 * scale)
         for layer_setting in bottleneck_params_list:
             t, c, n, s = layer_setting
+            i += 1
             input = self.invresi_blocks(
                 input=input,
                 in_c=in_c,
                 t=t,
                 c=int(c * scale),
                 n=n,
-                s=s, )
+                s=s,
+                name='conv' + str(i))
             in_c = int(c * scale)
-
+        #last_conv
         input = self.conv_bn_layer(
             input=input,
             num_filters=int(1280 * scale) if scale > 1.0 else 1280,
             filter_size=1,
             stride=1,
             padding=0,
-            if_act=True)
+            if_act=True,
+            name='conv9')
 
         input = fluid.layers.pool2d(
             input=input,
@@ -73,7 +80,8 @@ class MobileNetV2():
 
         output = fluid.layers.fc(input=input,
                                  size=class_dim,
-                                 param_attr=ParamAttr(initializer=MSRA()))
+                                 param_attr=ParamAttr(name='fc10_weights'),
+                                 bias_attr=ParamAttr(name='fc10_offset'))
         return output
 
     def conv_bn_layer(self,
@@ -84,8 +92,9 @@ class MobileNetV2():
                       padding,
                       channels=None,
                       num_groups=1,
-                      use_cudnn=True,
-                      if_act=True):
+                      if_act=True,
+                      name=None,
+                      use_cudnn=True):
         conv = fluid.layers.conv2d(
             input=input,
             num_filters=num_filters,
@@ -95,9 +104,15 @@ class MobileNetV2():
             groups=num_groups,
             act=None,
             use_cudnn=use_cudnn,
-            param_attr=ParamAttr(initializer=MSRA()),
+            param_attr=ParamAttr(name=name + '_weights'),
             bias_attr=False)
-        bn = fluid.layers.batch_norm(input=conv)
+        bn_name = name + '_bn'
+        bn = fluid.layers.batch_norm(
+            input=conv,
+            param_attr=ParamAttr(name=bn_name + "_scale"),
+            bias_attr=ParamAttr(name=bn_name + "_offset"),
+            moving_mean_name=bn_name + '_mean',
+            moving_variance_name=bn_name + '_variance')
         if if_act:
             return fluid.layers.relu6(bn)
         else:
@@ -106,10 +121,18 @@ class MobileNetV2():
     def shortcut(self, input, data_residual):
         return fluid.layers.elementwise_add(input, data_residual)
 
-    def inverted_residual_unit(self, input, num_in_filter, num_filters,
-                               ifshortcut, stride, filter_size, padding,
-                               expansion_factor):
+    def inverted_residual_unit(self,
+                               input,
+                               num_in_filter,
+                               num_filters,
+                               ifshortcut,
+                               stride,
+                               filter_size,
+                               padding,
+                               expansion_factor,
+                               name=None):
         num_expfilter = int(round(num_in_filter * expansion_factor))
+
         channel_expand = self.conv_bn_layer(
             input=input,
             num_filters=num_expfilter,
@@ -117,7 +140,9 @@ class MobileNetV2():
             stride=1,
             padding=0,
             num_groups=1,
-            if_act=True)
+            if_act=True,
+            name=name + '_expand')
+
         bottleneck_conv = self.conv_bn_layer(
             input=channel_expand,
             num_filters=num_expfilter,
@@ -126,7 +151,9 @@ class MobileNetV2():
             padding=padding,
             num_groups=num_expfilter,
             if_act=True,
+            name=name + '_dwise',
             use_cudnn=False)
+
         linear_out = self.conv_bn_layer(
             input=bottleneck_conv,
             num_filters=num_filters,
@@ -134,14 +161,15 @@ class MobileNetV2():
             stride=1,
             padding=0,
             num_groups=1,
-            if_act=False)
+            if_act=False,
+            name=name + '_linear')
         if ifshortcut:
             out = self.shortcut(input=input, data_residual=linear_out)
             return out
         else:
             return linear_out
 
-    def invresi_blocks(self, input, in_c, t, c, n, s):
+    def invresi_blocks(self, input, in_c, t, c, n, s, name=None):
         first_block = self.inverted_residual_unit(
             input=input,
             num_in_filter=in_c,
@@ -150,7 +178,8 @@ class MobileNetV2():
             stride=s,
             filter_size=3,
             padding=1,
-            expansion_factor=t)
+            expansion_factor=t,
+            name=name + '_1')
 
         last_residual_block = first_block
         last_c = c
@@ -164,5 +193,6 @@ class MobileNetV2():
                 stride=1,
                 filter_size=3,
                 padding=1,
-                expansion_factor=t)
+                expansion_factor=t,
+                name=name + '_' + str(i + 1))
         return last_residual_block
