+++ fluid/PaddleCV/image_classification/models/shufflenet_v2.py
@@ -52,7 +52,8 @@ class ShuffleNetV2():
             filter_size=3,
             num_filters=input_channel,
             padding=1,
-            stride=2)
+            stride=2,
+            name='stage1_conv')
         pool1 = fluid.layers.pool2d(
             input=conv1,
             pool_size=3,
@@ -70,30 +71,35 @@ class ShuffleNetV2():
                         input=conv,
                         num_filters=output_channel,
                         stride=2,
-                        benchmodel=2)
+                        benchmodel=2,
+                        name=str(idxstage + 2) + '_' + str(i + 1))
                 else:
                     conv = self.inverted_residual_unit(
                         input=conv,
                         num_filters=output_channel,
                         stride=1,
-                        benchmodel=1)
+                        benchmodel=1,
+                        name=str(idxstage + 2) + '_' + str(i + 1))
 
         conv_last = self.conv_bn_layer(
             input=conv,
             filter_size=1,
             num_filters=stage_out_channels[-1],
             padding=0,
-            stride=1)
+            stride=1,
+            name='conv5')
         pool_last = fluid.layers.pool2d(
             input=conv_last,
             pool_size=7,
-            pool_stride=7,
+            pool_stride=1,
             pool_padding=0,
             pool_type='avg')
 
         output = fluid.layers.fc(input=pool_last,
                                  size=class_dim,
-                                 param_attr=ParamAttr(initializer=MSRA()))
+                                 param_attr=ParamAttr(
+                                     initializer=MSRA(), name='fc6_weights'),
+                                 bias_attr=ParamAttr(name='fc6_offset'))
         return output
 
     def conv_bn_layer(self,
@@ -104,7 +110,9 @@ class ShuffleNetV2():
                       padding,
                       num_groups=1,
                       use_cudnn=True,
-                      if_act=True):
+                      if_act=True,
+                      name=None):
+        #         print(num_groups)
         conv = fluid.layers.conv2d(
             input=input,
             num_filters=num_filters,
@@ -114,12 +122,25 @@ class ShuffleNetV2():
             groups=num_groups,
             act=None,
             use_cudnn=use_cudnn,
-            param_attr=ParamAttr(initializer=MSRA()),
+            param_attr=ParamAttr(
+                initializer=MSRA(), name=name + '_weights'),
             bias_attr=False)
+        bn_name = name + '_bn'
         if if_act:
-            return fluid.layers.batch_norm(input=conv, act='relu')
+            return fluid.layers.batch_norm(
+                input=conv,
+                act='relu',
+                param_attr=ParamAttr(name=bn_name + "_scale"),
+                bias_attr=ParamAttr(name=bn_name + "_offset"),
+                moving_mean_name=bn_name + '_mean',
+                moving_variance_name=bn_name + '_variance')
         else:
-            return fluid.layers.batch_norm(input=conv)
+            return fluid.layers.batch_norm(
+                input=conv,
+                param_attr=ParamAttr(name=bn_name + "_scale"),
+                bias_attr=ParamAttr(name=bn_name + "_offset"),
+                moving_mean_name=bn_name + '_mean',
+                moving_variance_name=bn_name + '_variance')
 
     def channel_shuffle(self, x, groups):
         batchsize, num_channels, height, width = x.shape[0], x.shape[
@@ -138,7 +159,12 @@ class ShuffleNetV2():
 
         return x
 
-    def inverted_residual_unit(self, input, num_filters, stride, benchmodel):
+    def inverted_residual_unit(self,
+                               input,
+                               num_filters,
+                               stride,
+                               benchmodel,
+                               name=None):
         assert stride in [1, 2], \
             "supported stride are {} but your stride is {}".format([1,2], stride)
 
@@ -150,6 +176,8 @@ class ShuffleNetV2():
                 input,
                 num_or_sections=[input.shape[1] // 2, input.shape[1] // 2],
                 dim=1)
+            #             x1 = input[:, :(input.shape[1]//2), :, :]
+            #             x2 = input[:, (input.shape[1]//2):, :, :]
 
             conv_pw = self.conv_bn_layer(
                 input=x2,
@@ -158,7 +186,8 @@ class ShuffleNetV2():
                 stride=1,
                 padding=0,
                 num_groups=1,
-                if_act=True)
+                if_act=True,
+                name='stage_' + name + '_conv1')
 
             conv_dw = self.conv_bn_layer(
                 input=conv_pw,
@@ -167,7 +196,8 @@ class ShuffleNetV2():
                 stride=stride,
                 padding=1,
                 num_groups=oup_inc,
-                if_act=False)
+                if_act=False,
+                name='stage_' + name + '_conv2')
 
             conv_linear = self.conv_bn_layer(
                 input=conv_dw,
@@ -176,57 +206,63 @@ class ShuffleNetV2():
                 stride=1,
                 padding=0,
                 num_groups=1,
-                if_act=True)
+                if_act=True,
+                name='stage_' + name + '_conv3')
 
             out = fluid.layers.concat([x1, conv_linear], axis=1)
 
         else:
             #branch1
-            conv_dw = self.conv_bn_layer(
+            conv_dw_1 = self.conv_bn_layer(
                 input=input,
                 num_filters=inp,
                 filter_size=3,
                 stride=stride,
                 padding=1,
                 num_groups=inp,
-                if_act=False)
+                if_act=False,
+                name='stage_' + name + '_conv4')
 
             conv_linear_1 = self.conv_bn_layer(
-                input=conv_dw,
+                input=conv_dw_1,
                 num_filters=oup_inc,
                 filter_size=1,
                 stride=1,
                 padding=0,
                 num_groups=1,
-                if_act=True)
+                if_act=True,
+                name='stage_' + name + '_conv5')
 
             #branch2
-            conv_pw = self.conv_bn_layer(
+            conv_pw_2 = self.conv_bn_layer(
                 input=input,
                 num_filters=oup_inc,
                 filter_size=1,
                 stride=1,
                 padding=0,
                 num_groups=1,
-                if_act=True)
+                if_act=True,
+                name='stage_' + name + '_conv1')
 
-            conv_dw = self.conv_bn_layer(
-                input=conv_pw,
+            conv_dw_2 = self.conv_bn_layer(
+                input=conv_pw_2,
                 num_filters=oup_inc,
                 filter_size=3,
                 stride=stride,
                 padding=1,
                 num_groups=oup_inc,
-                if_act=False)
+                if_act=False,
+                name='stage_' + name + '_conv2')
 
             conv_linear_2 = self.conv_bn_layer(
-                input=conv_dw,
+                input=conv_dw_2,
                 num_filters=oup_inc,
                 filter_size=1,
                 stride=1,
                 padding=0,
                 num_groups=1,
-                if_act=True)
+                if_act=True,
+                name='stage_' + name + '_conv3')
             out = fluid.layers.concat([conv_linear_1, conv_linear_2], axis=1)
 
         return self.channel_shuffle(out, 2)
